+++ generators/app/templates/src/StarterKit/Framework/Exceptions/Constants.cs
@@ -1,0 +1,11 @@
+namespace StarterKit.Framework.Exceptions
+{
+  public class Defaults
+  {
+    internal class KeyNotFoundException
+    {
+      internal const string Title = "Key not found.";
+      internal const string Code = "KEYNFOUND001";
+    }
+  }
+}